Use excludeKey instead of includeKey to include expandSubWorkflows
Cromwell API appears to have a bug where it would not
expand subWorkflowMetadata when includeKey(s) are used.

Instead, we now have to explicitly enumerate all keys
that should be *excluded*.

We preserved the earlier implementation using includeKey
in a TODO comment, to be reverted to when Cromwell
fixes this.